NSFW Rewrite + Bug Fixes
Added an ID system to the NSFW commands. It is now possible to select an image to be sent by saying the corresponding number. e.g. $trap 67 would send the image 67.png in the cogs/NSFW/Traps directory. It automatically defaults to a random image, but I added this because someone requested it. On random images it also now says the ID of the image.

I also changed something with a few different commands in the hopes that it will prevent people from making the bot ping @everyone.

Also changed something with currency to hopefully fix saving money.

I fixed the missing prefix on 4 commands in $help too.